update: move class FeignClient
