--- src/main/kotlin/gogo/gogoshop/global/internal/participant/api/ParticipantApi.kt
@@ -1,7 +1,0 @@
-package gogo.gogoshop.global.internal.participant.api
-
-import gogo.gogoshop.global.internal.participant.stub.IsParticipantByStageIdAndStudentIdStub
-
-interface ParticipantApi {
-    fun isParticipantByStageIdAndStudentId(stageId: Long, studentId: Long): IsParticipantByStageIdAndStudentIdStub
-}
