--- src/main/kotlin/gogo/gogoshop/global/internal/participant/stub/ParticipantStub.kt
@@ -1,5 +1,0 @@
-package gogo.gogoshop.global.internal.participant.stub
-
-data class IsParticipantByStageIdAndStudentIdStub(
-    val isParticipant: Boolean,
-)
