--- src/main/kotlin/gogo/gogoshop/global/internal/point/api/PointApiImpl.kt
@@ -1,12 +1,0 @@
-package gogo.gogoshop.global.internal.point.api
-
-import gogo.gogoshop.global.feign.client.PointClient
-import org.springframework.stereotype.Component
-
-@Component
-class PointApiImpl(
-    private val pointClient: PointClient
-): PointApi {
-    override fun queryPointByStageIdAndStudentId(stageId: Long, studentId: Long) =
-        pointClient.queryPointByStageIdAndStudentId(stageId, studentId)
-}
