--- src/main/kotlin/gogo/gogoshop/global/internal/point/stub/PointStub.kt
@@ -1,5 +1,0 @@
-package gogo.gogoshop.global.internal.point.stub
-
-data class PointStub(
-    val point: Long
-)
