+++ src/main/kotlin/gogo/gogoshop/global/internal/point/stub/StageStub.kt
@@ -1,0 +1,9 @@
+package gogo.gogoshop.global.internal.point.stub
+
+data class StageStub(
+    val point: Long
+)
+
+data class IsParticipantByStageIdAndStudentIdStub(
+    val isParticipant: Boolean,
+)
